Moved out content-scoped js-instrumentation to separate file
